feat: add admin password login and redesign Vue admin UI

+++ server/routes/api/accounts/[id].delete.ts
@@ -1,11 +1,11 @@
 import { defineEventHandler, getRouterParam } from "h3"
-import { requireProxyAuth } from "../../../utils/auth"
+import { requireAdminAuth } from "../../../utils/admin-auth"
 import { sendManagementError } from "../../../utils/errors"
 import { deleteAccount } from "../../../utils/store"
 
 export default defineEventHandler(async (event) => {
   try {
-    await requireProxyAuth(event)
+    await requireAdminAuth(event)
     const id = getRouterParam(event, "id")
     if (!id) throw new Error("Account id is required")
     await deleteAccount(id)
